Separate extensions and types into different files
